Move variable declaration out of the loop, make generic tag text more generic
As neither change.type, change.old.tags nor change.neu.tags is modified in
the loop, the value of tags will always be the same, so this can be moved
outside the loop.
Instead of using the hard-coded 'a way' text for the fallback tag text,
take the actual element type from the change, which is more generic in case
the element is something which is not a way.